move stubbing code to seperate file
